% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgelist.R
\name{as.dgelist}
\alias{as.dgelist}
\alias{as.dgelist.default}
\alias{as.dgelist.fpkm_counts}
\alias{as.dgelist.raw_counts}
\title{Convert to a \code{DGEList} object}
\usage{
as.dgelist(x, ...)

\method{as.dgelist}{default}(x, ...)

\method{as.dgelist}{fpkm_counts}(x, group = NULL, ...)

\method{as.dgelist}{raw_counts}(x, group = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{fpkm_counts} or \code{raw_counts}.}

\item{...}{Additional arguments specific to method.}

\item{group}{column name in 'x' corresponding to the group each sample 
belongs to.}
}
\value{
An object of class \code{DGEList}.
}
\description{
At the moment, these objects can be converted to a 
\code{DGEList} object directly by doing \code{as.dgelist(x)}.
\itemize{
\item \code{fpkm_counts} -- object created by calling 
\code{ganalyse::gather_counts} on a \code{fpkm} object.

\item \code{raw_counts}  -- object created by calling 
\code{ganalyse::gather_counts} on a \code{raw} object. 
}
}
\details{
The \code{as.dgelist} method is a simple wrapper for converting 
\code{raw_counts} and \code{fpkm_counts} objects to \code{DGEList} object. 
No additional transformations or filtering is done, i.e., it assumes that 
the counts are already filtered and (log) transformed as required. Use the 
arguments in \code{gather_counts()} method if necessary to prepare the 
counts as necessary.
}
\examples{
path = system.file("tests", package="ganalyse")

# ----- fpkm ----- # 
fpkm_path = file.path(path, "fpkm", "annotation.txt")
fpkm_obj = rnaseq(fpkm_path, format="fpkm", experiment="sample")
fpkm_counts = gather_counts(fpkm_obj, by="gene-id", log_base=2L)
(as.dgelist(fpkm_counts))

# ----- raw ----- # 
raw_path = file.path(path, "raw", "annotation.txt")
raw_obj = rnaseq(raw_path, format="raw", experiment="sample")
raw_counts = gather_counts(raw_obj, by="gene-id", threshold=1L)
(as.dgelist(raw_counts))

}
\seealso{
\code{\link{rnaseq}}, \code{\link{gather_counts}} 
\code{\link{show_counts}} \code{\link{limma_dge}} \code{\link{edger_dge}} 
\code{\link{construct_design}} \code{\link{construct_contrasts}} 
\code{\link{write_dge}} \code{\link{as.eset}}
\code{\link{volcano_plot}} \code{\link{density_plot}}
}

