% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esets.R
\name{as.eset}
\alias{as.eset}
\alias{as.eset.DGEList}
\alias{as.eset.default}
\alias{as.eset.fpkm_counts}
\alias{as.eset.raw_counts}
\title{Convert objects to an \code{ExpressionSet} object}
\usage{
as.eset(x, ...)

\method{as.eset}{default}(x, ...)

\method{as.eset}{DGEList}(x, groups = NULL, ...)

\method{as.eset}{fpkm_counts}(x, groups = NULL, ...)

\method{as.eset}{raw_counts}(x, groups = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{fpkm_format}, \code{raw_format} or 
\code{DGEList}.}

\item{...}{Additional arguments specific to methods.}

\item{groups}{columns in 'x' corresponding to the group each sample 
belongs to.}
}
\value{
An object of class \code{ExpressionSet}.
}
\description{
At the moment, these objects can be converted to an 
\code{ExpressionSet} object directly by doing \code{as.eset(x)}.
\itemize{
\item \code{fpkm_counts} -- object created by calling 
\code{ganalyse::gather_counts} on a \code{fpkm_format} object.

\item \code{raw_counts}  -- object created by calling 
\code{ganalyse::gather_counts} on a \code{raw_format} object. 

\item \code{DGEList} -- object created by calling \code{edgeR::DGEList} 
function.
}
}
\examples{
path = system.file("tests", package="ganalyse")

# ----- fpkm ----- # 
fpkm_path = file.path(path, "fpkm", "annotation.txt")
fpkm_obj = rnaseq(fpkm_path, format="fpkm", experiment="sample")
fpkm_counts = gather_counts(fpkm_obj, by="gene-id", log_base=2L)
(as.eset(fpkm_counts))

# ----- raw ----- # 
raw_path = file.path(path, "raw", "annotation.txt")
raw_obj = rnaseq(raw_path, format="raw", experiment="sample")
raw_counts = gather_counts(raw_obj, by="gene-id", threshold=1L)
(as.eset(raw_counts))

}
\seealso{
\code{\link{rnaseq}}, \code{\link{gather_counts}} 
\code{\link{show_counts}} \code{\link{limma_dge}} \code{\link{edger_dge}} 
\code{\link{construct_design}} \code{\link{construct_contrasts}} 
\code{\link{write_dge}} \code{\link{as.dgelist}}
\code{\link{volcano_plot}} \code{\link{density_plot}}
}

