% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esets-map.R
\name{as.eset_map}
\alias{as.eset_map}
\alias{as.eset_map.default}
\alias{as.eset_map.fpkm_counts}
\alias{as.eset_map.raw_counts}
\title{Detailed \code{ExpressionSet} object}
\usage{
as.eset_map(x, entrez_map = NULL, reference_group = "", ...)

\method{as.eset_map}{default}(x, entrez_map = NULL, ...)

\method{as.eset_map}{fpkm_counts}(x, entrez_map = NULL, genome_annotation,
  gene_name_col, reference_group = "", ...)

\method{as.eset_map}{raw_counts}(x, entrez_map = NULL, genome_annotation,
  gene_name_col, reference_group = "", ...)
}
\arguments{
\item{x}{An object of class \code{fpkm_counts}, \code{raw_counts} 
or \code{DGEList}.}

\item{entrez_map}{Full path to file that contains mapping of gene names to 
ENTREZ ids along with corresponding gene ids . It must be a three column 
\code{data.table}, with column names \code{gene_name}, \code{entrez_id} 
and \code{gene_id} respectively. 

If the argument is \code{NULL} (default), the column \code{ENTREZID} in 
object \code{pData(featureData(<eset_obj>))} will be set to \code{NA}.}

\item{reference_group}{A character vector of length=1 indicating the 
reference sample.}

\item{...}{Additional arguments specific to methods.}

\item{genome_annotation}{An object of class \code{gtf} or \code{gff}.}

\item{gene_name_col}{A character vector of length=1 indicating the name of 
the column corresponding to gene names in \code{genome_annotation}}
}
\value{
An object of class \code{ExpressionSet} tuned for MAP.
}
\description{
\code{as.eset_map()} converts an object of class 
\code{fpkm_counts} or \code{raw_counts} to an \code{ExpressionSet} object, 
with a lot of other details that are normally required for DGE analysis, 
but might be necessary at times. 

This function is still in development, hence not exported.
}
\examples{
\dontrun{
  # my_gtf is obtained using gread::read_format.
  eset_obj = as.eset_map(x, genome_annotation = my_gtf, 
                   gene_name_col = "gene_name")
}
}

