% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density-plot.R
\name{density_plot}
\alias{density-plot}
\alias{density_plot}
\alias{density_plot.default}
\alias{density_plot.fpkm_counts}
\alias{density_plot.raw_counts}
\title{Density plots of count data facetted by groups}
\usage{
density_plot(x, interactive = FALSE, title, groups, facet_cols = 1L,
  log = FALSE, filename, height = 8L, width = 12L, ...)

\method{density_plot}{default}(x, interactive = FALSE, title, groups,
  facet_cols = 1L, log = FALSE, filename, height = 8L, width = 12L, ...)

\method{density_plot}{fpkm_counts}(x, interactive = FALSE, title, groups,
  facet_cols = 1L, log = FALSE, filename, height = 8L, width = 12L, ...)

\method{density_plot}{raw_counts}(x, interactive = FALSE, title, groups,
  facet_cols = 1L, log = FALSE, filename, height = 8L, width = 12L, ...)
}
\arguments{
\item{x}{An object of class \code{dge} or a list of objects of class 
\code{dge}.}

\item{interactive}{Default is \code{FALSE}. If \code{TRUE}, uses 
\code{ggplotly::ggplotly()} to return an interactive plot.}

\item{title}{Plot title.}

\item{groups}{Columns indicating those samples that needs to be 
selected together while facetting.}

\item{facet_cols}{Number of columns to provide to \code{facet_wrap}.}

\item{log}{Should values be log-transformed? Default is \code{FALSE}.}

\item{filename}{Default is to plot to screen. If a file name 
is provided, the plot is saved to file and the plot object will be 
returned. The type of graphic device is auto-detected from file extension 
(using \code{ggsave}).}

\item{height}{Height of the plot, default is 8 inches.}

\item{width}{Width of the plot, default is 12 inches.}

\item{...}{Optional arguments to pass to \code{ggsave} call.}
}
\value{
A \code{ggplot2} or \code{plotly} object containing the 
density plot of input \code{fpkm_counts} or \code{raw_counts} object.
}
\description{
\code{density_plot} function takes an object of class 
\code{fpkm_counts} or \code{raw_counts} and returns density plots of all 
samples facetted by the groups specified.
}
\examples{
path = system.file("tests", package="ganalyse")

# ----- raw ----- # 
raw_path = file.path(path, "raw", "annotation.txt")
raw_obj = rnaseq(raw_path, format="raw", experiment="sample")
raw_counts = gather_counts(raw_obj, by="gene-id", threshold=1L)

# DE genes between treatA and control
density_plot(raw_counts, interactive=FALSE, groups="condition") # ggplot2
density_plot(raw_counts, interactive=FALSE, groups="condition", log=TRUE)
density_plot(raw_counts, interactive=FALSE, 
        groups="condition", title="Raw count density")
density_plot(raw_counts, interactive=TRUE, 
        groups="condition", log=TRUE) # ggplotly
\dontrun{
# write to file and return plot obj
density_plot(raw_counts, filename="tmp.png")
}
}
\seealso{
\code{\link{rnaseq}}, \code{\link{gather_counts}} 
\code{\link{show_counts}} \code{\link{limma_dge}} \code{\link{edger_dge}}
\code{\link{construct_design}} \code{\link{construct_contrasts}} 
\code{\link{write_dge}} \code{\link{as.dgelist}} \code{\link{as.eset}}
}

