% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.R
\name{rnaseq}
\alias{rnaseq}
\title{Create an RNASeq experiment object}
\usage{
rnaseq(sample_info, experiment = "example", format = c("fpkm", "raw"),
  verbose = FALSE)
}
\arguments{
\item{sample_info}{Complete path to file that contains \code{sample} names, 
columns that identify the \code{groups} each \code{sample} belongs to 
and path to the file that contains \emph{read counts} for each sample. 
See \code{details} section for more.}

\item{experiment}{Name of the experiment. Default is "example".}

\item{format}{Is the input data \code{fpkm} or \code{raw} counts?}

\item{verbose}{Logical. Default is \code{FALSE}. If \code{TRUE}, sends 
useful status messages to the console.}
}
\value{
An object of class \code{fpkm} or \code{raw}. It 
inherits from \code{data.table}.
}
\description{
Loads all sample info related to an experiment on to an object 
of class \code{fpkm} or \code{raw}.
}
\details{
The file provided to \code{sample_info} (say 
\code{annotation.txt}) \bold{must} have columns named \code{sample} 
(corresponding to sample names) and \code{path} (containing the file path 
to \code{raw} or \code{fpkm} counts). The path for each sample can either 
be a complete path or just the filename (with extension). If it is just a 
file name, it is assumed to be in the same path as \code{annotation.txt}.

The other columns usually identify which groups and/or treatments each 
sample belongs to. This allows easy construction of design matrix to 
describe treatment conditions later on, with the help of 
\code{\link{construct_design}}. These columns are optional at this point 
and can be added manually later on as well.

The easiest way would be to place \code{annotation.txt} and the counts in 
the folder.
}
\examples{
path = system.file("tests", package="ganalyse")

# ----- fpkm ----- # 
fpkm_path = file.path(path, "fpkm", "annotation.txt")
(fpkm_obj = rnaseq(fpkm_path, format="fpkm", experiment="sample"))
class(fpkm_obj)

# ----- raw ----- # 
raw_path = file.path(path, "raw", "annotation.txt")
(raw_obj = rnaseq(raw_path, format="raw", experiment="sample"))
class(raw_obj)
}
\seealso{
\code{\link{gather_counts}} \code{\link{show_counts}}
\code{\link{limma_dge}} \code{\link{edger_dge}} 
\code{\link{construct_design}} \code{\link{construct_contrasts}}
\code{\link{write_dge}} \code{\link{as.dgelist}} \code{\link{as.eset}}
\code{\link{volcano_plot}} \code{\link{density_plot}}
}

