% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show-counts.R
\name{show_counts}
\alias{show_counts}
\alias{show_counts.default}
\alias{show_counts.fpkm_counts}
\alias{show_counts.raw_counts}
\title{Display all counts in a rectangular format}
\usage{
show_counts(x)

\method{show_counts}{default}(x)

\method{show_counts}{fpkm_counts}(x)

\method{show_counts}{raw_counts}(x)
}
\arguments{
\item{x}{An object of class \code{fpkm_counts} or \code{raw_counts}. See 
\code{?gather_counts}.}
}
\value{
An object of class \code{data.table} containing the feature and 
associated read counts for each sample in a tabular format.
}
\description{
This function is for interactive purposes to inspect the 
read counts. The first column corresponds to feature id and all other 
columns correspond to read counts for each sample.
}
\examples{
path = system.file("tests", package="ganalyse")

# ----- fpkm ----- # 
fpkm_path = file.path(path, "fpkm", "annotation.txt")
fpkm_obj = rnaseq(fpkm_path, format="fpkm", experiment="sample")
fpkm_counts = gather_counts(fpkm_obj, by="gene-id", log_base=2L)
head(show_counts(fpkm_counts))

# ----- raw ----- # 
raw_path = file.path(path, "raw", "annotation.txt")
raw_obj = rnaseq(raw_path, format="raw", experiment="sample")
raw_counts = gather_counts(raw_obj, by="gene-id", threshold=1L)
head(show_counts(raw_counts))
}
\seealso{
\code{\link{rnaseq}} \code{\link{limma_dge}} 
\code{\link{edger_dge}} \code{\link{as.eset}} \code{\link{gather_counts}}
\code{\link{construct_design}} \code{\link{construct_contrasts}}
\code{\link{volcano_plot}} \code{\link{write_dge}} 
\code{\link{density_plot}}
}

