% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpkm-dge.R
\name{write_dge}
\alias{write_dge}
\title{Write top table results to file.}
\usage{
write_dge(x, filename)
}
\arguments{
\item{x}{A list of top tables}

\item{filename}{A character vector of same length as \code{x} containing 
filenames corresponding to each element in \code{x}.}
}
\value{
It returns a \code{NULL} value if writing file to disk 
completes successfully.
}
\description{
This is a convenience function for writing multiple top 
tables to files. It takes a list of top tables and corresponding file 
names and writes each table to that file.
}
\examples{
path = system.file("tests", package="ganalyse")

# ----- raw ----- # 
raw_path = file.path(path, "raw", "annotation.txt")
raw_obj = rnaseq(raw_path, format="raw", experiment="sample")
raw_counts = gather_counts(raw_obj, by="gene-id", threshold=1L)
raw_design = construct_design(raw_counts, ~ 0 + condition)
raw_contrasts = construct_contrasts(
                     design = raw_design, 
                     treatA.vs.control = conditiontreatA-conditioncontrol, 
                     treatB.vs.control = conditiontreatB-conditioncontrol
                 )
# DE genes between treatA and control
ans = edger_dge(raw_counts, design=raw_design, 
             contrast=raw_contrasts[, "treatA.vs.control"])
\dontrun{
write_dge(ans, "treatA_vs_control.tsv")
}
}
\seealso{
\code{\link{rnaseq}} \code{\link{gather_counts}} 
\code{\link{show_counts}} \code{\link{limma_dge}} \code{\link{edger_dge}} 
\code{\link{construct_design}} \code{\link{construct_contrasts}}
\code{\link{as.dgelist}} \code{\link{as.eset}} 
\code{\link{volcano_plot}}
}

